mqtt_listener/requirements.txt
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
python-dotenv==1.0.1
pyyaml==6.0.1
